move num_bigint_impl to own module
